backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.2
celery==5.3.6
redis==5.0.3
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
PyPDF2==3.0.1
pdfminer.six==20231228
python-magic==0.4.27
gunicorn==21.2.0
django-storages==1.14.2
boto3==1.34.34
